refactor(hopper): extract field validation from form.service.ts
Extract validateFieldValue() (pure per-type validation logic) from
form.service.ts into form-validation.service.ts. The new module is
pure — no DB dependencies, takes field definition + value, returns
errors.

form.service.ts: 913 → 698 lines. validateFormData() stays as the
coordination point (loads form from DB, iterates fields, delegates
per-field validation to the extracted function).

Tests: 27 new tests in form-validation.service.spec.ts (direct pure
function tests), 6 coordination tests remain in form.service.spec.ts.
